server/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
tensorflow==2.15.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.9
